requirements.txt
opencv-python==4.8.1.78
tqdm==4.65.0
pytorch==2.0.1
matplotlib==3.7.2
numpy==1.26.0
scipy==1.11.3
scikit-image==0.20.0
scikit-learn==1.3.1==
imageio==2.31.4
plotly==5.9.0